+++ x/allocation/types/query.pb.gw.go
@@ -133,6 +133,10 @@ func local_request_Query_Option_0(ctx context.Context, marshaler runtime.Marshal
 
 }
 
+var (
+	filter_Query_Options_0 = &utilities.DoubleArray{Encoding: map[string]int{"stream": 0}, Base: []int{1, 1, 0}, Check: []int{0, 1, 2}}
+)
+
 func request_Query_Options_0(ctx context.Context, marshaler runtime.Marshaler, client QueryClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
 	var protoReq QueryOptionsRequest
 	var metadata runtime.ServerMetadata
@@ -157,6 +161,13 @@ func request_Query_Options_0(ctx context.Context, marshaler runtime.Marshaler, c
 	}
 
 	protoReq.Stream = StreamId(e)
+
+	if err := req.ParseForm(); err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
+	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_Query_Options_0); err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
 
 	msg, err := client.Options(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
 	return msg, metadata, err
@@ -187,6 +198,13 @@ func local_request_Query_Options_0(ctx context.Context, marshaler runtime.Marsha
 	}
 
 	protoReq.Stream = StreamId(e)
+
+	if err := req.ParseForm(); err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
+	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_Query_Options_0); err != nil {
+		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
+	}
 
 	msg, err := server.Options(ctx, &protoReq)
 	return msg, metadata, err